requirements.txt
torch>=2.3.0
sentence-transformers==3.0.1
transformers==4.40.0
numpy==1.26.4
python-dotenv>=1.0.1
pathspec>=0.12.1